\name{yalda-package}
\alias{yalda-package}
\docType{package}
\title{yalda}
\description{
  TODO
}
\details{
\tabular{ll}{
  Package: \tab rth\cr
  Type: \tab Package\cr
  License: \tab Apache 2.0\cr
}
}
\references{
  \url{http://github.com/wrathematics/yalda}
}
\author{
  Drew Schmidt
}
\keyword{Package}
